create temp file for test purposes
