+++ svcs/amigo/amigo_test.go
@@ -10,14 +10,22 @@ import (
 	"github.com/aau-network-security/haaukins/svcs/amigo"
 	"google.golang.org/grpc"
 	"io/ioutil"
+	"log"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"testing"
 )
 
 func TestVerifyFlag(t *testing.T) {
-	skey := "testing"
+	// temporary events directory for NewStoreEvent
+	tmp, err := ioutil.TempDir("", "")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer os.RemoveAll(tmp)
 
+	skey := "testing"
 	dialer, close := store.CreateTestServer()
 	defer close()
 
@@ -41,7 +49,7 @@ func TestVerifyFlag(t *testing.T) {
 		StartedAt:      nil,
 		FinishExpected: nil,
 		FinishedAt:     nil,
-	}, "events", client)
+	},tmp ,client)
 
 	var chal = store.FlagConfig{
 			Tag:         "test",
